+++ CHANGELOG.md
@@ -1,4 +1,13 @@
-# Improvements
+# 2.0.0
+. Migrated to React 18
+. Uses create react app utility
+. Removes enzyme
+. Uses react test library
+. Fixes test
+. Typescript 4
+. Migrate rollup to latest
+
+# Improvements 1.2.17
 
 . Change warning color
 . optimized build